refactor: reorganize database functions and remove dotenv usage from API files
