+++ database/__init__.py
@@ -1,0 +1,10 @@
+from .MongoDB import (
+    get_user,
+    create_local_user,
+    create_google_user,
+    update_user,
+    update_password
+)
+
+
+__all__ = ['get_user', 'create_local_user', 'create_google_user', 'update_user', 'update_password']
